feat(hc-login): Update login flow to run in webbrowser.
This is to support the new singlekey-id.com authentication flow.

It works by setting the URL handler for the hcauth:// page that the login
flow redirects to, and passing that through a UNIX socket to the hc-login script.

This may be improved by using a IP server to support authentication
on remote machines, but that may not be needed since you can create config.json
on your local machine and copy that to the remote that needs to run hc2mqtt.